% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{assignFlyway}
\alias{assignFlyway}
\title{Assign flyway}
\usage{
assignFlyway(data)
}
\arguments{
\item{data}{Estimation data}
}
\description{
The internal \code{assignFlyway} function assigns flyway abbreviations and
flyway numbers to estimation data. To assign flyways to sea duck or brant
estimation data, see \code{\link{assignFlywaySDBR}}.
}
\seealso{
Other estimation functions: 
\code{\link{assignFlywaySDBR}()},
\code{\link{assignMgmtUnit}()},
\code{\link{calcAllStats}()},
\code{\link{calcFreq}()},
\code{\link{calcSumMeanVar}()},
\code{\link{calcVar}()},
\code{\link{flywayFinal}()},
\code{\link{flywayFinalSDBR}()},
\code{\link{flywayTotals}()},
\code{\link{harvestEstimates}()},
\code{\link{hunterEstimates}()},
\code{\link{mgmtUnitFinal}()},
\code{\link{mgmtUnitSelect}()},
\code{\link{mgmtUnitTotals}()},
\code{\link{speciesEstimate}()},
\code{\link{stateSelect}()},
\code{\link{stateTotals}()},
\code{\link{usFinal}()},
\code{\link{usFinalDV}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{estimation functions}
