% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{assignFlywaySDBR}
\alias{assignFlywaySDBR}
\title{Assign flyway to sea duck and brant estimates}
\usage{
assignFlywaySDBR(data)
}
\arguments{
\item{data}{Estimation data}
}
\description{
The internal \code{assignFlywaySDBR} function assigns flyway abbreviations
and flyway numbers to sea duck and brant estimation data.
}
\seealso{
Other estimation functions: 
\code{\link{assignFlyway}()},
\code{\link{assignMgmtUnit}()},
\code{\link{calcAllStats}()},
\code{\link{calcFreq}()},
\code{\link{calcSumMeanVar}()},
\code{\link{calcVar}()},
\code{\link{flywayFinal}()},
\code{\link{flywayFinalSDBR}()},
\code{\link{flywayTotals}()},
\code{\link{harvestEstimates}()},
\code{\link{hunterEstimates}()},
\code{\link{mgmtUnitFinal}()},
\code{\link{mgmtUnitSelect}()},
\code{\link{mgmtUnitTotals}()},
\code{\link{speciesEstimate}()},
\code{\link{stateSelect}()},
\code{\link{stateTotals}()},
\code{\link{usFinal}()},
\code{\link{usFinalDV}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{estimation functions}
