% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{audit}
\alias{audit}
\title{Check survey data}
\usage{
audit(daily_check, season_check)
}
\arguments{
\item{daily_check}{Tibble output from a \code{checkDaily} function}

\item{season_check}{Tibble output from a \code{checkSeason} function}
}
\description{
Internal function used in \code{\link{surveyCheck}} that returns corrections
and audits. Used for cranes, snipe, coot, rails, gallinules, and woodcock.
For doves, use \code{\link{auditDV}}. For waterfowl, use
\code{\link{checkWF}}.
}
\seealso{
Other checking functions: 
\code{\link{auditDV}()},
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{checkWF}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{joinSeason}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other crane functions: 
\code{\link{checkDailySCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()}

Other SCRG functions: 
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other woodcock functions: 
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{SCRG functions}
\concept{checking functions}
\concept{crane functions}
\concept{woodcock functions}
