% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxPlotWWDO}
\alias{boxPlotWWDO}
\title{Plot White-winged Dove survey data}
\usage{
boxPlotWWDO(season_df, variable, y_limits = "auto")
}
\arguments{
\item{season_df}{Season data tibble}

\item{variable}{Variable to plot; should be one of "days_hunted", "retrieved"
or "unretrieved"}

\item{y_limits}{If "auto", \code{ggplot2::coord_cartesian} \code{ylim}
supplied as \code{c(0, 50)} for days_hunted and retrieved, \code{c(0, 20)}
for unretrieved; otherwise, provide custom limits.}
}
\description{
Create a box plot of non-zero White-winged Dove days hunted, retrieved, or
unretrieved values.
}
\seealso{
Other dove functions: 
\code{\link{auditDV}()},
\code{\link{checkDailyDV}()},
\code{\link{checkSeasonDV}()},
\code{\link{convertWWDO}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{dove functions}
