% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{calcAllStats}
\alias{calcAllStats}
\title{Calculate all statistics}
\usage{
calcAllStats(licenseSumMeanVar, licenseFreq, spp_counts, type)
}
\arguments{
\item{licenseSumMeanVar}{Product of \code{calcSumMeanVar}}

\item{licenseFreq}{\code{calcFreq}}

\item{spp_counts}{Count data tibble}

\item{type}{Type of calculation to run, may be one of: 'Ducks', 'Geese',
'Brant', 'SeaDucks', 'MODO', 'WWDO', 'SACR', 'Woodcock', 'Snipe', 'Coots',
'Rails', 'Gallinules', 'BTPI'.}
}
\description{
The internal \code{calcAllStats} function combines all stratum-level hunter
statistics using the products of \code{\link{calcSumMeanVar}} and
\code{\link{calcFreq}}.
}
\seealso{
Other estimation functions: 
\code{\link{calcFreq}()},
\code{\link{calcSumMeanVar}()},
\code{\link{calcVar}()},
\code{\link{estimate}()},
\code{\link{summarizeByState}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{estimation functions}
