% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{calcFreq}
\alias{calcFreq}
\title{Calculate frequency}
\usage{
calcFreq(totals_df, type)
}
\arguments{
\item{totals_df}{Data tibble}

\item{type}{Type of calculation to run, may be one of: 'Ducks', 'Geese',
'Brant', 'SeaDucks', 'MODO', 'WWDO', 'SACR', 'Woodcock', 'Snipe', 'Coots',
'Rails', 'Gallinules', 'BTPI'.}
}
\description{
The internal \code{calcFreq} function calculates the number of hunters by
state and stratum.
}
\seealso{
Other estimation functions: 
\code{\link{calcAllStats}()},
\code{\link{calcSumMeanVar}()},
\code{\link{calcVar}()},
\code{\link{estimate}()},
\code{\link{summarizeByState}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{estimation functions}
