% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{calcSumMeanVar}
\alias{calcSumMeanVar}
\title{Calculate sum, mean, and variance}
\usage{
calcSumMeanVar(totals_df, type)
}
\arguments{
\item{totals_df}{Data tibble}

\item{type}{Type of calculation to run, may be one of: 'Ducks', 'Geese',
'Brant', 'SeaDucks', 'MODO', 'WWDO', 'SACR', 'Woodcock', 'Snipe', 'Coots',
'Rails', 'Gallinules', 'BTPI'.}
}
\description{
The internal \code{calcSumMeanVar} function the calculates sum, mean, and
variance of data, given a species group.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
