% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkDaily}
\alias{checkDaily}
\title{Check daily data}
\usage{
checkDaily(dailies_df, maxbag_df, species)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{maxbag_df}{Reference data tibble}

\item{species}{Species abbreviation, may be one of: 'WF', 'DV', 'SCRG', 'WK', or 'CR'.}
}
\description{
Internal function used in \code{\link{editCheck}}.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
