% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkDailySCRGWKCR}
\alias{checkDailySCRGWKCR}
\title{Check daily data}
\usage{
checkDailySCRGWKCR(dailies_df, maxbag_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{maxbag_df}{Reference data tibble}
}
\description{
Internal function to check daily survey data for SC, RG, WK, CR. Used in
\code{\link{surveyCheck}}.
}
\seealso{
Other checking functions: 
\code{\link{audit}()},
\code{\link{auditDV}()},
\code{\link{checkDailyDV}()},
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{checkWF}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{joinSeason}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other daily data helpers: 
\code{\link{checkDailyDV}()},
\code{\link{checkDailyWF}()},
\code{\link{convertSDBR}()},
\code{\link{convertWWDO}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{partyHuntFinder}()}

Other crane functions: 
\code{\link{audit}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()}

Other SCRG functions: 
\code{\link{audit}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other woodcock functions: 
\code{\link{audit}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{SCRG functions}
\concept{checking functions}
\concept{crane functions}
\concept{daily data helpers}
\concept{woodcock functions}
