% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSeasonCR}
\alias{checkSeasonCR}
\title{Check crane season data}
\usage{
checkSeasonCR(season_df, maxbag_df)
}
\arguments{
\item{season_df}{Season data tibble}

\item{maxbag_df}{Reference data tibble}
}
\description{
Internal function to check season survey data for cranes.
}
\seealso{
Other checking functions: 
\code{\link{audit}()},
\code{\link{auditDV}()},
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWK}()},
\code{\link{checkWF}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()},
\code{\link{tooManyDaysHuntedCR}()}

Other season data helpers: 
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWK}()},
\code{\link{convertSDBR}()},
\code{\link{convertWWDO}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{tooManyDaysHunted}()},
\code{\link{tooManyDaysHuntedCR}()}

Other crane functions: 
\code{\link{audit}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHuntedCR}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{checking functions}
\concept{crane functions}
\concept{season data helpers}
