% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSeasonWFSCRGWK}
\alias{checkSeasonWFSCRGWK}
\title{Check season data}
\usage{
checkSeasonWFSCRGWK(season_df, maxbag_df, day_limit)
}
\arguments{
\item{season_df}{Season data tibble}

\item{maxbag_df}{Reference data tibble}

\item{day_limit}{Day limit for species being checked}
}
\description{
Internal function to check season survey data for waterfowl, snipe, coot,
rails, gallinules, and woodcock.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
