% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSeasonWFSCRGWKCR}
\alias{checkSeasonWFSCRGWKCR}
\title{Check season data}
\usage{
checkSeasonWFSCRGWKCR(season_df, maxbag_df, maxbag_raw)
}
\arguments{
\item{season_df}{Season data tibble}

\item{maxbag_df}{Reference data tibble}

\item{maxbag_raw}{Raw reference data tibble}
}
\description{
Internal function to check season survey data for waterfowl, snipe, coot,
rails, gallinules, and woodcock. Used in \code{\link{surveyCheck}}.
}
\seealso{
Other checking functions: 
\code{\link{audit}()},
\code{\link{auditDV}()},
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonDV}()},
\code{\link{checkWF}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{joinSeason}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other season data helpers: 
\code{\link{checkSeasonDV}()},
\code{\link{convertSDBR}()},
\code{\link{convertWWDO}()},
\code{\link{joinSeason}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{tooManyDaysHunted}()}

Other SCRG functions: 
\code{\link{audit}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other waterfowl functions: 
\code{\link{checkDailyWF}()},
\code{\link{checkWF}()},
\code{\link{convertSDBR}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other woodcock functions: 
\code{\link{audit}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{SCRG functions}
\concept{checking functions}
\concept{season data helpers}
\concept{waterfowl functions}
\concept{woodcock functions}
