% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertBrantToGeese}
\alias{convertBrantToGeese}
\title{Convert brant to geese for brant records from non-brant states}
\usage{
convertBrantToGeese(dailies_df, season_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{season_df}{Season data tibble}
}
\description{
Internal function used in \code{\link{editCheck}}.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
