% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertSDBR}
\alias{convertSDBR}
\title{Convert sea ducks to ducks and brant to geese}
\usage{
convertSDBR(daily_corrected, season_precorrected)
}
\arguments{
\item{daily_corrected}{Corrected daily data tibble}

\item{season_precorrected}{Corrected season data tibble}
}
\description{
Internal function used in \code{\link{checkWF}}. If harvest is reported from
a state without a sea duck season and/or brant season, convert the sea duck
harvest to duck harvest and brant harvest to goose harvest.
}
\seealso{
Other conversion functions: 
\code{\link{convertBrantToGeese}()},
\code{\link{convertSeaDuckToDuck}()},
\code{\link{convertWWDO}()},
\code{\link{summarizeWWDO}()}

Other daily data helpers: 
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{convertWWDO}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{partyHuntFinder}()}

Other season data helpers: 
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{convertWWDO}()},
\code{\link{joinSeason}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{tooManyDaysHunted}()}

Other waterfowl functions: 
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{checkWF}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{conversion functions}
\concept{daily data helpers}
\concept{season data helpers}
\concept{waterfowl functions}
