% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertSeaDuckToDuck}
\alias{convertSeaDuckToDuck}
\title{Convert sea ducks to ducks for sea duck records from non-sea duck states}
\usage{
convertSeaDuckToDuck(dailies_df, season_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{season_df}{Season data tibble}
}
\description{
Internal function used in \code{\link{convertSDBR}}.
}
\seealso{
Other conversion functions: 
\code{\link{convertBrantToGeese}()},
\code{\link{convertSDBR}()},
\code{\link{convertWWDO}()},
\code{\link{summarizeWWDO}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{conversion functions}
