% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertWWDO}
\alias{convertWWDO}
\title{Find season or daily WWDO harvest effort in non-WWDO states and convert to
zero}
\usage{
convertWWDO(data_df, type, summary = FALSE)
}
\arguments{
\item{data_df}{Season or daily data tibble}

\item{type}{"season" or "daily"}

\item{summary}{Whether a summary of WWDO errors should be returned; TRUE or
FALSE (default is FALSE)}
}
\description{
Internal function used in \code{\link{checkDailyDV}} and
\code{\link{checkSeasonDV}}. Find any harvest of White-winged Dove in states
without expected White-winged Dove harvest and in edge states that exceed the
designated limit; for these cases, change White-winged Dove
\code{days_hunted}, \code{retrieved}, and \code{unretrieved} values to
\code{0}.
}
\seealso{
Other conversion functions: 
\code{\link{convertBrantToGeese}()},
\code{\link{convertSDBR}()},
\code{\link{convertSeaDuckToDuck}()},
\code{\link{summarizeWWDO}()}

Other daily data helpers: 
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{convertSDBR}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{partyHuntFinder}()}

Other season data helpers: 
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{convertSDBR}()},
\code{\link{joinSeason}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{tooManyDaysHunted}()}

Other dove functions: 
\code{\link{auditDV}()},
\code{\link{boxPlotWWDO}()},
\code{\link{checkDailyDV}()},
\code{\link{checkSeasonDV}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{conversion functions}
\concept{daily data helpers}
\concept{dove functions}
\concept{season data helpers}
