% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{dailyAudit}
\alias{dailyAudit}
\title{Audit daily data}
\usage{
dailyAudit(daily_checked, daily_error_ids)
}
\arguments{
\item{daily_checked}{Daily data tibble checked by \code{\link{checkDaily}}}

\item{daily_error_ids}{Survey IDs of records with errors in the daily data}
}
\description{
Internal function used in \code{\link{editCheck}}.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
