% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{dailyCorrect}
\alias{dailyCorrect}
\title{Correct daily data}
\usage{
dailyCorrect(daily_checked, daily_error_ids, species)
}
\arguments{
\item{daily_checked}{Daily data tibble checked by \code{\link{checkDaily}}}

\item{daily_error_ids}{Survey IDs of records with errors in the daily data}

\item{species}{Species abbreviation, may be one of: 'WF', 'DV', 'SCRG', 'WK', or 'CR'.}
}
\description{
Internal function used in \code{\link{editCheck}}.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
