% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorIDs.R
\name{dailyErrorIDs}
\alias{dailyErrorIDs}
\title{Pull daily data surveyIDs for records that contain errors}
\usage{
dailyErrorIDs(daily_checked, species)
}
\arguments{
\item{daily_checked}{Daily data tibble checked by \code{\link{checkDaily}}}

\item{species}{Species abbreviation, may be one of: 'WF', 'DV', 'SCRG', 'WK', or 'CR'.}
}
\description{
Internal function used in \code{\link{editCheck}}.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
