% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{dailyOverBagDKSD}
\alias{dailyOverBagDKSD}
\title{Duck and sea duck over bag}
\usage{
dailyOverBagDKSD(dailies_df, maxbag_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{maxbag_df}{Reference data tibble}
}
\description{
Internal function used in \code{\link{checkDailyWF}}. Used to find daily
overbag for ducks and sea ducks. Creates field \code{error2}.
}
\seealso{
Other checking functions: 
\code{\link{audit}()},
\code{\link{auditDV}()},
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{checkWF}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{joinSeason}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other daily data helpers: 
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{convertSDBR}()},
\code{\link{convertWWDO}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{partyHuntFinder}()}

Other error assigning functions: 
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{tooManyDaysHunted}()}

Other waterfowl functions: 
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{checkWF}()},
\code{\link{convertSDBR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{checking functions}
\concept{daily data helpers}
\concept{error assigning functions}
\concept{waterfowl functions}
