% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{editCheck}
\alias{editCheck}
\title{Check survey data}
\usage{
editCheck(dailies_df, season_df, maxbag_df, species)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{season_df}{Season data tibble}

\item{maxbag_df}{Reference data tibble}

\item{species}{Species abbreviation, may be one of: 'WF', 'DV', 'SCRG', 'WK', or 'CR'.}
}
\description{
Check season and daily data for a given species and return corrections and audits.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
