% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{fails}
\alias{fails}
\title{Fail data}
\usage{
fails(season_df, dailies_df, n_states)
}
\arguments{
\item{season_df}{Season data tibble}

\item{dailies_df}{Daily data tibble}

\item{n_states}{Number of expected states in the data}
}
\description{
Internal function used to repeat validation on season and daily survey data.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
