% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failspp}
\alias{failspp}
\title{Fail species}
\usage{
failspp(species)
}
\arguments{
\item{species}{Species abbreviation, may be one of: 'WF', 'DV', 'SCRG', 'WK',
or 'CR'.}
}
\description{
Internal function used to validate the \code{species} provided.
}
\seealso{
Other failure functions: 
\code{\link{failNADaysHunted}()},
\code{\link{failNARetrieved}()},
\code{\link{failStateCount}()},
\code{\link{failSurveyStates}()},
\code{\link{failtype}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
