% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failSurveyStates}
\alias{failSurveyStates}
\title{Fail survey states}
\usage{
failSurveyStates(data, type)
}
\arguments{
\item{data}{Survey data}

\item{type}{Type of calculation to run, may be one of: 'Ducks', 'Geese',
'Brant', 'SeaDucks', 'MODO', 'WWDO', 'SACR', 'Woodcock', 'Snipe', 'Coots',
'Rails', 'Gallinules', 'BTPI'.}
}
\description{
Internal function used to validate the \code{state} values in the data,
depending on the \code{type} provided.
}
\seealso{
Other failure functions: 
\code{\link{failNADaysHunted}()},
\code{\link{failNARetrieved}()},
\code{\link{failStateCount}()},
\code{\link{failspp}()},
\code{\link{failtype}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
