% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{flywayFinalSDBR}
\alias{flywayFinalSDBR}
\title{Final flyway estimates for sea duck and brant}
\usage{
flywayFinalSDBR(data)
}
\arguments{
\item{data}{Estimation data}
}
\description{
The internal \code{flywayFinalSDBR} function produces the final flyway
harvest estimates for sea ducks and brant.
}
\seealso{
Other estimation functions: 
\code{\link{assignFlyway}()},
\code{\link{assignFlywaySDBR}()},
\code{\link{assignMgmtUnit}()},
\code{\link{calcAllStats}()},
\code{\link{calcFreq}()},
\code{\link{calcSumMeanVar}()},
\code{\link{calcVar}()},
\code{\link{flywayFinal}()},
\code{\link{flywayTotals}()},
\code{\link{harvestEstimates}()},
\code{\link{hunterEstimates}()},
\code{\link{mgmtUnitFinal}()},
\code{\link{mgmtUnitSelect}()},
\code{\link{mgmtUnitTotals}()},
\code{\link{speciesEstimate}()},
\code{\link{stateSelect}()},
\code{\link{stateTotals}()},
\code{\link{usFinal}()},
\code{\link{usFinalDV}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{estimation functions}
