% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party.R
\name{getPartySize}
\alias{getPartySize}
\title{Get party size}
\usage{
getPartySize(dailies_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}
}
\description{
Internal function used in \code{\link{partyHuntFinder}}. Pull out obvious
group sizes from strings and put them in a new 'party_size' field.
}
\seealso{
Other party hunt functions: 
\code{\link{partyHuntFinder}()},
\code{\link{recalcPartyBag}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{party hunt functions}
