% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSestimate-package.R
\docType{package}
\name{HSestimate-package}
\alias{HSestimate}
\alias{HSestimate-package}
\title{HSestimate: Calculate Harvest Estimates from the USFWS Migratory Bird Harvest Survey Data in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions designed for the USFWS Migratory Bird Program to generate harvest estimates using Harvest Survey response data.
}
\author{
\strong{Maintainer}: Abby Walter \email{abby_walter@fws.gov} (\href{https://orcid.org/0000-0003-4017-5596}{ORCID})

}
\keyword{internal}
