% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{naDaysHunted}
\alias{naDaysHunted}
\title{Find records with NA days hunted}
\usage{
naDaysHunted(season_df)
}
\arguments{
\item{season_df}{Season data tibble}
}
\description{
Internal function used in \code{\link{checkSeasonWFSCRGWKCR}} and
\code{\link{checkSeasonDV}}. Finds season records with \code{NA} values in
\code{days_hunted} field. Creates field \code{error1}.
}
\seealso{
Other checking functions: 
\code{\link{audit}()},
\code{\link{auditDV}()},
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{checkWF}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{joinSeason}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other season data helpers: 
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{convertSDBR}()},
\code{\link{convertWWDO}()},
\code{\link{joinSeason}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{tooManyDaysHunted}()}

Other error assigning functions: 
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{tooManyDaysHunted}()}

Other crane functions: 
\code{\link{audit}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()}

Other dove functions: 
\code{\link{auditDV}()},
\code{\link{boxPlotWWDO}()},
\code{\link{checkDailyDV}()},
\code{\link{checkSeasonDV}()},
\code{\link{convertWWDO}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other SCRG functions: 
\code{\link{audit}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other waterfowl functions: 
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{checkWF}()},
\code{\link{convertSDBR}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other woodcock functions: 
\code{\link{audit}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{SCRG functions}
\concept{checking functions}
\concept{crane functions}
\concept{dove functions}
\concept{error assigning functions}
\concept{season data helpers}
\concept{waterfowl functions}
\concept{woodcock functions}
