% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party.R
\name{partyHuntFinder}
\alias{partyHuntFinder}
\title{Party hunt finder}
\usage{
partyHuntFinder(dailies_df, maxbag_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{maxbag_df}{Reference data tibble}
}
\description{
Parse comment strings to check if a daily record was a party hunt and
calculate revised retrieved field if value exceeds max bag. Used in
\code{\link{checkDailySCRGWKCR}}, \code{\link{checkDailyDV}}, and
\code{\link{checkDailyWF}}. Creates field \code{error1}.
}
\seealso{
Other party hunt functions: 
\code{\link{getPartySize}()},
\code{\link{recalcPartyBag}()}

Other daily data helpers: 
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{convertSDBR}()},
\code{\link{convertWWDO}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()}

Other error assigning functions: 
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{tooManyDaysHunted}()}

Other crane functions: 
\code{\link{audit}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()}

Other dove functions: 
\code{\link{auditDV}()},
\code{\link{boxPlotWWDO}()},
\code{\link{checkDailyDV}()},
\code{\link{checkSeasonDV}()},
\code{\link{convertWWDO}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other SCRG functions: 
\code{\link{audit}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other waterfowl functions: 
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{checkWF}()},
\code{\link{convertSDBR}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other woodcock functions: 
\code{\link{audit}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{SCRG functions}
\concept{crane functions}
\concept{daily data helpers}
\concept{dove functions}
\concept{error assigning functions}
\concept{party hunt functions}
\concept{waterfowl functions}
\concept{woodcock functions}
