% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{seasonCorrect}
\alias{seasonCorrect}
\title{Correct season data}
\usage{
seasonCorrect(season_checked, season_error_ids, daily_error_ids, species)
}
\arguments{
\item{season_checked}{Season data tibble checked by \code{\link{checkSeason}}}

\item{season_error_ids}{Survey IDs of records with errors in the season data}

\item{daily_error_ids}{Survey IDs of records with errors in the daily data}

\item{species}{Species abbreviation, may be one of: 'WF', 'DV', 'SCRG', 'WK', or 'CR'.}
}
\description{
Internal function used in \code{\link{editCheck}}.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
