% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorIDs.R
\name{seasonErrorIDs}
\alias{seasonErrorIDs}
\title{Pull season data surveyIDs for records that contain errors}
\usage{
seasonErrorIDs(season_checked)
}
\arguments{
\item{season_checked}{Season data tibble checked by \code{\link{checkSeason}}}
}
\description{
Internal function used in \code{\link{editCheck}}.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
