% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{speciesEstimate}
\alias{speciesEstimate}
\title{Estimate harvest for a species group}
\usage{
speciesEstimate(totals_df, spp_counts, type)
}
\arguments{
\item{totals_df}{Data tibble}

\item{spp_counts}{Count data tibble}

\item{type}{Type of calculation to run, may be one of: 'Ducks', 'Geese',
'Brant', 'SeaDucks', 'MODO', 'WWDO', 'SACR', 'Woodcock', 'Snipe', 'Coots',
'Rails', 'Gallinules', 'BTPI'.}
}
\description{
Estimate species group harvest; sum the weighted stratum means.
}
\seealso{
Other estimation functions: 
\code{\link{assignFlyway}()},
\code{\link{assignFlywaySDBR}()},
\code{\link{assignMgmtUnit}()},
\code{\link{calcAllStats}()},
\code{\link{calcFreq}()},
\code{\link{calcSumMeanVar}()},
\code{\link{calcVar}()},
\code{\link{flywayFinal}()},
\code{\link{flywayFinalSDBR}()},
\code{\link{flywayTotals}()},
\code{\link{harvestEstimates}()},
\code{\link{hunterEstimates}()},
\code{\link{mgmtUnitFinal}()},
\code{\link{mgmtUnitSelect}()},
\code{\link{mgmtUnitTotals}()},
\code{\link{stateSelect}()},
\code{\link{stateTotals}()},
\code{\link{usFinal}()},
\code{\link{usFinalDV}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{estimation functions}
