% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{summarizeByState}
\alias{summarizeByState}
\title{Summarize estimates by state}
\usage{
summarizeByState(strat_est_df, type)
}
\arguments{
\item{strat_est_df}{Stratum estimates tibble}

\item{type}{Type of calculation to run, may be one of: 'Ducks', 'Geese',
'Brant', 'SeaDucks', 'MODO', 'WWDO', 'SACR', 'Woodcock', 'Snipe', 'Coots',
'Rails', 'Gallinules', 'BTPI'.}
}
\description{
The internal \code{summarizeByState} function summarizes harvest estimates by
state.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
