% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{summarizeWWDO}
\alias{summarizeWWDO}
\title{Summarize WWDO errors}
\usage{
summarizeWWDO(errors_df, type)
}
\arguments{
\item{errors_df}{Season or daily data tibble}

\item{type}{"season" or "daily"}
}
\description{
Internal function used in \code{\link{convertWWDO}}.
}
\seealso{
Other conversion functions: 
\code{\link{convertBrantToGeese}()},
\code{\link{convertSDBR}()},
\code{\link{convertSeaDuckToDuck}()},
\code{\link{convertWWDO}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{conversion functions}
