% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{tooManyDaysHunted}
\alias{tooManyDaysHunted}
\title{Find records with too many days hunted}
\usage{
tooManyDaysHunted(season_df, maxbag_raw)
}
\arguments{
\item{season_df}{Season data tibble}

\item{maxbag_raw}{Raw reference data tibble}
}
\description{
Internal function used in \code{\link{checkSeasonWFSCRGWKCR}} and
\code{\link{checkSeasonDV}}. Finds season records with too many days hunted.
Creates field \code{error2}.
}
\seealso{
Other checking functions: 
\code{\link{audit}()},
\code{\link{auditDV}()},
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{checkWF}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{joinSeason}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()}

Other season data helpers: 
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{convertSDBR}()},
\code{\link{convertWWDO}()},
\code{\link{joinSeason}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()}

Other error assigning functions: 
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()}

Other dove functions: 
\code{\link{auditDV}()},
\code{\link{boxPlotWWDO}()},
\code{\link{checkDailyDV}()},
\code{\link{checkSeasonDV}()},
\code{\link{convertWWDO}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()}

Other SCRG functions: 
\code{\link{audit}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()}

Other waterfowl functions: 
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{checkWF}()},
\code{\link{convertSDBR}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()}

Other woodcock functions: 
\code{\link{audit}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkSeasonWFSCRGWKCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{SCRG functions}
\concept{checking functions}
\concept{dove functions}
\concept{error assigning functions}
\concept{season data helpers}
\concept{waterfowl functions}
\concept{woodcock functions}
