% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{usFinal}
\alias{usFinal}
\title{Final United States estimates}
\usage{
usFinal(data)
}
\arguments{
\item{data}{Estimation data}
}
\description{
The internal \code{usFinal} function produces the final national harvest
estimates. For the final estimates for doves, see \code{\link{usFinalDV}}.
}
\seealso{
Other estimation functions: 
\code{\link{assignFlyway}()},
\code{\link{assignFlywaySDBR}()},
\code{\link{assignMgmtUnit}()},
\code{\link{calcAllStats}()},
\code{\link{calcFreq}()},
\code{\link{calcSumMeanVar}()},
\code{\link{calcVar}()},
\code{\link{flywayFinal}()},
\code{\link{flywayFinalSDBR}()},
\code{\link{flywayTotals}()},
\code{\link{harvestEstimates}()},
\code{\link{hunterEstimates}()},
\code{\link{mgmtUnitFinal}()},
\code{\link{mgmtUnitSelect}()},
\code{\link{mgmtUnitTotals}()},
\code{\link{speciesEstimate}()},
\code{\link{stateSelect}()},
\code{\link{stateTotals}()},
\code{\link{usFinalDV}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{estimation functions}
