% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangleDates}
\alias{wrangleDates}
\title{Wrangle max bag}
\usage{
wrangleDates(ref_data)
}
\arguments{
\item{ref_data}{Reference data}
}
\description{
Wrangle the raw max bag reference table.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
