% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangleDayLimits}
\alias{wrangleDayLimits}
\title{Wrangle day limits}
\usage{
wrangleDayLimits(maxbag_raw)
}
\arguments{
\item{maxbag_raw}{Raw max bag tibble}
}
\description{
Wrangle the raw max bag reference table to get hunting season day limits for
all states and species.
}
\seealso{
Other wrangling functions: 
\code{\link{wrangleDates}()},
\code{\link{wrangleMaxBag}()},
\code{\link{wrangleRef}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{wrangling functions}
