% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangleMaxBag}
\alias{wrangleMaxBag}
\title{Wrangle max bag}
\usage{
wrangleMaxBag(ref_data)
}
\arguments{
\item{ref_data}{Reference data}
}
\description{
Wrangle the raw max bag reference table to get bag limits for all states and
species.
}
\seealso{
Other wrangling functions: 
\code{\link{wrangleDates}()},
\code{\link{wrangleDayLimits}()},
\code{\link{wrangleRef}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{wrangling functions}
