% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangleRef}
\alias{wrangleRef}
\title{Wrangle reference data}
\usage{
wrangleRef(ref_data)
}
\arguments{
\item{ref_data}{Reference data}
}
\description{
Internal function which does initial data cleaning in
\code{\link{wrangleMaxBag}} and \code{\link{wrangleDates}}.
}
\seealso{
Other wrangling functions: 
\code{\link{wrangleDates}()},
\code{\link{wrangleDayLimits}()},
\code{\link{wrangleMaxBag}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{wrangling functions}
