% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{assignMgmtUnit}
\alias{assignMgmtUnit}
\title{Assign management units}
\usage{
assignMgmtUnit(data, management_units)
}
\arguments{
\item{data}{Estimation data}

\item{management_units}{Tibble of management units}
}
\description{
The internal \code{assignMgmtUnit} function assigns management unit abbreviations
and management unit numbers to estimation data.
}
\seealso{
Other estimation functions: 
\code{\link{assignFlyway}()},
\code{\link{assignFlywaySDBR}()},
\code{\link{calcAllStats}()},
\code{\link{calcFreq}()},
\code{\link{calcSumMeanVar}()},
\code{\link{calcVar}()},
\code{\link{flywayFinal}()},
\code{\link{flywayFinalSDBR}()},
\code{\link{flywayTotals}()},
\code{\link{harvestEstimates}()},
\code{\link{hunterEstimates}()},
\code{\link{mgmtUnitFinal}()},
\code{\link{mgmtUnitSelect}()},
\code{\link{mgmtUnitTotals}()},
\code{\link{speciesEstimate}()},
\code{\link{stateSelect}()},
\code{\link{stateTotals}()},
\code{\link{usFinal}()},
\code{\link{usFinalDV}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{estimation functions}
