% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{audit}
\alias{audit}
\title{Check survey data}
\usage{
audit(daily_check, season_check)
}
\arguments{
\item{daily_check}{Tibble output from a \code{checkDaily} function}

\item{season_check}{Tibble output from a \code{checkSeason} function}
}
\description{
Internal function that checks season and daily data to return corrections and
audits.
}
\seealso{
Other checking functions: 
\code{\link{auditDV}()},
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonCR}()},
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWK}()},
\code{\link{checkWF}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()},
\code{\link{tooManyDaysHuntedCR}()}

Other crane functions: 
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkSeasonCR}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHuntedCR}()}

Other SCRG functions: 
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkSeasonWFSCRGWK}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}

Other woodcock functions: 
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkSeasonWFSCRGWK}()},
\code{\link{dailyOverBag}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{SCRG functions}
\concept{checking functions}
\concept{crane functions}
\concept{woodcock functions}
