% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{auditDV}
\alias{auditDV}
\title{Check dove survey data}
\usage{
auditDV(daily_check, season_check)
}
\arguments{
\item{daily_check}{Tibble output from a \code{checkDaily} function}

\item{season_check}{Tibble output from a \code{checkSeason} function}
}
\description{
Internal function that checks season and daily data for doves to return
corrections and audits.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
