% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{calcAllStats}
\alias{calcAllStats}
\title{Calculate all stats}
\usage{
calcAllStats(licenseSumMeanVar, licenseFreq, spp_counts, type)
}
\arguments{
\item{licenseSumMeanVar}{Product of \code{calcSumMeanVar}}

\item{licenseFreq}{\code{calcFreq}}

\item{spp_counts}{Count data tibble}

\item{type}{Type of calculation to run, may be one of: 'Ducks', 'Geese',
'Brant', 'SeaDucks', 'MODO', 'WWDO', 'SACR', 'Woodcock', 'Snipe', 'Coots',
'Rails', 'Gallinules', 'BTPI'.}
}
\description{
The internal \code{calcAllStats} function calculates all stats.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
