% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{calcVar}
\alias{calcVar}
\title{Calculate variance}
\usage{
calcVar(alldata_df, type)
}
\arguments{
\item{alldata_df}{Data tibble}

\item{type}{Type of calculation to run, may be one of: 'Ducks', 'Geese',
'Brant', 'SeaDucks', 'MODO', 'WWDO', 'SACR', 'Woodcock', 'Snipe', 'Coots',
'Rails', 'Gallinules', 'BTPI'.}
}
\description{
The internal \code{calcFreq} function calculates variance and other values.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
