% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{calcVar}
\alias{calcVar}
\title{Calculate variance}
\usage{
calcVar(alldata_df, type)
}
\arguments{
\item{alldata_df}{Product of \code{calcAllStats}}

\item{type}{Type of calculation to run, may be one of: 'Ducks', 'Geese',
'Brant', 'SeaDucks', 'MODO', 'WWDO', 'SACR', 'Woodcock', 'Snipe', 'Coots',
'Rails', 'Gallinules', 'BTPI'.}
}
\description{
The internal \code{calcVar} function calculates variance and other values
using the product of \code{\link{calcAllStats}}.
}
\seealso{
Other estimation functions: 
\code{\link{assignFlyway}()},
\code{\link{assignFlywaySDBR}()},
\code{\link{assignMgmtUnit}()},
\code{\link{calcAllStats}()},
\code{\link{calcFreq}()},
\code{\link{calcSumMeanVar}()},
\code{\link{flywayFinal}()},
\code{\link{flywayFinalSDBR}()},
\code{\link{flywayTotals}()},
\code{\link{harvestEstimates}()},
\code{\link{hunterEstimates}()},
\code{\link{mgmtUnitFinal}()},
\code{\link{mgmtUnitSelect}()},
\code{\link{mgmtUnitTotals}()},
\code{\link{speciesEstimate}()},
\code{\link{stateSelect}()},
\code{\link{stateTotals}()},
\code{\link{usFinal}()},
\code{\link{usFinalDV}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{estimation functions}
