% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkDailySCRGWKCR}
\alias{checkDailySCRGWKCR}
\title{Check daily data}
\usage{
checkDailySCRGWKCR(dailies_df, maxbag_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{maxbag_df}{Reference data tibble}
}
\description{
Internal function to check daily survey data for SC, RG, WK, CR.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
