% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkDailyWF}
\alias{checkDailyWF}
\title{Check waterfowl daily data}
\usage{
checkDailyWF(dailies_df, maxbag_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{maxbag_df}{Reference data tibble}
}
\description{
Internal function to check daily survey data for waterfowl.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
