% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSeasonDV}
\alias{checkSeasonDV}
\title{Check dove season data}
\usage{
checkSeasonDV(season_df, maxbag_df)
}
\arguments{
\item{season_df}{Season data tibble}

\item{maxbag_df}{Reference data tibble}
}
\description{
Internal function to check season survey data for doves.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
