% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkWF}
\alias{checkWF}
\title{Check waterfowl survey data}
\usage{
checkWF(dailies_df, season_df, maxbag_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{season_df}{Season data tibble}

\item{maxbag_df}{Reference data tibble}
}
\description{
Internal function that checks season and daily data for waterfowl and returns
corrections and audits.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
