% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertSDBR}
\alias{convertSDBR}
\title{Convert sea ducks to ducks and brant to geese}
\usage{
convertSDBR(daily_corrected, season_precorrected)
}
\arguments{
\item{daily_corrected}{Daily data corrected by the \code{dailyCorrect} family
of functions}

\item{season_precorrected}{Season data pre-corrected by the
\code{seasonCorrect} family of functions}
}
\description{
Internal function used in \code{\link{surveyCheck}}.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
