% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertSDBR}
\alias{convertSDBR}
\title{Convert sea ducks to ducks and brant to geese}
\usage{
convertSDBR(daily_corrected, season_precorrected)
}
\arguments{
\item{daily_corrected}{Daily data corrected by the \code{dailyCorrect} family
of functions}

\item{season_precorrected}{Season data pre-corrected by the
\code{seasonCorrect} family of functions}
}
\description{
Internal function used in \code{\link{surveyCheck}}.
}
\seealso{
Other conversion functions: 
\code{\link{convertBrantToGeese}()},
\code{\link{convertSeaDuckToDuck}()},
\code{\link{convertWWDO}()},
\code{\link{summarizeWWDO}()}

Other daily data helpers: 
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{convertWWDO}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{partyHuntFinder}()}

Other season data helpers: 
\code{\link{checkSeasonCR}()},
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWK}()},
\code{\link{convertWWDO}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{tooManyDaysHunted}()},
\code{\link{tooManyDaysHuntedCR}()}

Other waterfowl functions: 
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonWFSCRGWK}()},
\code{\link{checkWF}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{conversion functions}
\concept{daily data helpers}
\concept{season data helpers}
\concept{waterfowl functions}
