% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertSeaDuckToDuck}
\alias{convertSeaDuckToDuck}
\title{Convert sea ducks to ducks for sea duck records from non-sea duck states}
\usage{
convertSeaDuckToDuck(dailies_df, season_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{season_df}{Season data tibble}
}
\description{
Internal function used in \code{\link{editCheck}}.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
