% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{dailyOverBagDKSD}
\alias{dailyOverBagDKSD}
\title{Duck and sea duck over bag}
\usage{
dailyOverBagDKSD(dailies_df, maxbag_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{maxbag_df}{Reference data tibble}
}
\description{
Internal function used in the \code{checkDaily} family of functions. Used to
find daily overbag for ducks and sea ducks.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
