% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overbag.R
\name{dailyOverBagWWMO}
\alias{dailyOverBagWWMO}
\title{WWDO and MODO over bag}
\usage{
dailyOverBagWWMO(dailies_df, maxbag_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{maxbag_df}{Reference data tibble}
}
\description{
Internal function used in \code{\link{checkDaily}}.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
