% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failNADaysHunted}
\alias{failNADaysHunted}
\title{Fail if NA values in days_hunted field}
\usage{
failNADaysHunted(season_df)
}
\arguments{
\item{season_df}{Season data tibble}
}
\description{
Internal function that fails if NA values are found in the \code{days_hunted}
field of season data.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
