% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failNADaysHunted}
\alias{failNADaysHunted}
\title{Fail if NA values in days_hunted field}
\usage{
failNADaysHunted(season_df)
}
\arguments{
\item{season_df}{Season data tibble}
}
\description{
Internal function that fails if \code{NA} values are found in the
\code{days_hunted} field of season data.
}
\seealso{
Other failure functions: 
\code{\link{failNARetrieved}()},
\code{\link{failStateCount}()},
\code{\link{failspp}()},
\code{\link{failtype}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
