% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failNARetrieved}
\alias{failNARetrieved}
\title{Fail if NA values in retrieved field}
\usage{
failNARetrieved(data)
}
\arguments{
\item{data}{Season or daily data tibble}
}
\description{
Internal function that fails if NA values are found in the \code{retrieved}
field of season and/or daily data.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
