% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failspp}
\alias{failspp}
\title{Fail species}
\usage{
failspp(species)
}
\arguments{
\item{species}{Species code string, may be one of: 'WF', 'DV', 'SCRG', 'WK', or 'CR'.}
}
\description{
Internal function used to repeat species code validation.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
