% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failtype}
\alias{failtype}
\title{Fail type}
\usage{
failtype(type)
}
\arguments{
\item{type}{Type of estimation being calculated, may be one of: 'Ducks',
'Geese', 'Brant', 'SeaDucks', 'MODO', 'WWDO', 'SACR', 'Woodcock', 'Snipe',
'Coots', 'Rails', 'Gallinules', 'BTPI'.}
}
\description{
Internal function used to validate the \code{type} provided, which aids in
harvest estimation calculation.
}
\seealso{
Other failure functions: 
\code{\link{failNADaysHunted}()},
\code{\link{failNARetrieved}()},
\code{\link{failStateCount}()},
\code{\link{failspp}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
