% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failtype}
\alias{failtype}
\title{Fail type}
\usage{
failtype(type)
}
\arguments{
\item{type}{Type of estimation being calculated, may be one of: 'Ducks',
'Geese', 'Brant', 'SeaDucks', 'MODO', 'WWDO', 'SACR', 'Woodcock', 'Snipe',
'Coots', 'Rails', 'Gallinules', 'BTPI'.}
}
\description{
Internal function used to validate calculation type.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
