% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{joinSeason}
\alias{joinSeason}
\title{Join maxbag to season data}
\usage{
joinSeason(season_df, maxbag_df)
}
\arguments{
\item{season_df}{Season data tibble}

\item{maxbag_df}{Reference data tibble}
}
\description{
Internal function to join in maxbag fields to season data. Used in
\code{\link{checkSeasonWFSCRGWK}}, \code{\link{checkSeasonDV}}, and
\code{\link{checkSeasonCR}}.
}
\seealso{
Other checking functions: 
\code{\link{audit}()},
\code{\link{auditDV}()},
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonCR}()},
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWK}()},
\code{\link{checkWF}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()},
\code{\link{tooManyDaysHuntedCR}()}

Other season data helpers: 
\code{\link{checkSeasonCR}()},
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWK}()},
\code{\link{convertSDBR}()},
\code{\link{convertWWDO}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{seasonOverBagDV}()},
\code{\link{tooManyDaysHunted}()},
\code{\link{tooManyDaysHuntedCR}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{checking functions}
\concept{season data helpers}
