% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{naDaysHunted}
\alias{naDaysHunted}
\title{Find records with NA days hunted}
\usage{
naDaysHunted(season_df)
}
\arguments{
\item{season_df}{Season data tibble}
}
\description{
Internal function used in \code{\link{checkSeason}}. Find season records with NA in days_hunted field.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
