% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party.R
\name{partyHuntFinder}
\alias{partyHuntFinder}
\title{Party hunt finder}
\usage{
partyHuntFinder(dailies_df, maxbag_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}

\item{maxbag_df}{Reference data tibble}
}
\description{
Parse comment strings to check if a daily record was a party hunt and calculate revised retrieved field if value exceeds max bag.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
