% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party.R
\name{recalcPartyBag}
\alias{recalcPartyBag}
\title{Recalculate party bag}
\usage{
recalcPartyBag(dailies_df)
}
\arguments{
\item{dailies_df}{Daily data tibble}
}
\description{
Internal function used in \code{\link{partyHuntFinder}}. Recalculate retrieved value for party hunt over bags by dividing retrieved by party size.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
