% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{seasonDNH}
\alias{seasonDNH}
\title{Find season did-not-hunts}
\usage{
seasonDNH(season_df)
}
\arguments{
\item{season_df}{Season data tibble}
}
\description{
Internal function used in the \code{checkSeason} family of functions. Find
season records with a sum of \code{0} for \code{days_hunted} and sum greater
than \code{0} for \code{retrieved}.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
