% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{seasonDNH}
\alias{seasonDNH}
\title{Find season did-not-hunts}
\usage{
seasonDNH(season_df)
}
\arguments{
\item{season_df}{Season data tibble}
}
\description{
Internal function used in \code{\link{checkSeason}}. Find season records with a sum of 0 for days_hunted and sum > 0 for retrieved.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
