% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overbag.R
\name{seasonOverBag}
\alias{seasonOverBag}
\title{Season over bag}
\usage{
seasonOverBag(season_df, species)
}
\arguments{
\item{season_df}{Season data tibble}

\item{species}{Species abbreviation, may be one of: 'WF', 'DV', 'SCRG', 'WK', or 'CR'.}
}
\description{
Internal function used in \code{\link{checkSeason}}.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
