% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{seasonOverBagDV}
\alias{seasonOverBagDV}
\title{Dove season over bag}
\usage{
seasonOverBagDV(dvtotals_df)
}
\arguments{
\item{dvtotals_df}{Season data tibble}
}
\description{
Internal function used in the \code{checkSeason} family of functions. Flag
records with an average overbag in season totals for doves more than the
overbag tolerance.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
