% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{seasonOverBagDV}
\alias{seasonOverBagDV}
\title{Dove season over bag}
\usage{
seasonOverBagDV(dvtotals_df)
}
\arguments{
\item{dvtotals_df}{Season data tibble}
}
\description{
Internal function used in \code{\link{checkSeasonDV}}. Flags records with an
average overbag in season totals for doves more than the overbag tolerance.
Creates field \code{error3}. For species other than doves, see
\code{\link{seasonOverBag}}.
}
\seealso{
Other checking functions: 
\code{\link{audit}()},
\code{\link{auditDV}()},
\code{\link{checkDailyDV}()},
\code{\link{checkDailySCRGWKCR}()},
\code{\link{checkDailyWF}()},
\code{\link{checkSeasonCR}()},
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWK}()},
\code{\link{checkWF}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()},
\code{\link{tooManyDaysHuntedCR}()}

Other season data helpers: 
\code{\link{checkSeasonCR}()},
\code{\link{checkSeasonDV}()},
\code{\link{checkSeasonWFSCRGWK}()},
\code{\link{convertSDBR}()},
\code{\link{convertWWDO}()},
\code{\link{naDaysHunted}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{tooManyDaysHunted}()},
\code{\link{tooManyDaysHuntedCR}()}

Other error assigning functions: 
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagDKSD}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{seasonOverBag}()},
\code{\link{tooManyDaysHunted}()},
\code{\link{tooManyDaysHuntedCR}()}

Other dove functions: 
\code{\link{auditDV}()},
\code{\link{boxPlotWWDO}()},
\code{\link{checkDailyDV}()},
\code{\link{checkSeasonDV}()},
\code{\link{convertWWDO}()},
\code{\link{dailyOverBag}()},
\code{\link{dailyOverBagWWMO}()},
\code{\link{naDaysHunted}()},
\code{\link{partyHuntFinder}()},
\code{\link{seasonDNH}()},
\code{\link{surveyCheck}()},
\code{\link{tooManyDaysHunted}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{checking functions}
\concept{dove functions}
\concept{error assigning functions}
\concept{season data helpers}
