% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{summarizeByState}
\alias{summarizeByState}
\title{Summarize estimates by state}
\usage{
summarizeByState(strat_est_df, type)
}
\arguments{
\item{strat_est_df}{Stratum estimates tibble, the product of \code{calcVar}}

\item{type}{Type of calculation to run, may be one of: 'Ducks', 'Geese',
'Brant', 'SeaDucks', 'MODO', 'WWDO', 'SACR', 'Woodcock', 'Snipe', 'Coots',
'Rails', 'Gallinules', 'BTPI'.}
}
\description{
The internal \code{summarizeByState} function summarizes harvest estimates by
state using the product of \code{\link{calcVar}}.
}
\seealso{
Other estimation functions: 
\code{\link{calcAllStats}()},
\code{\link{calcFreq}()},
\code{\link{calcSumMeanVar}()},
\code{\link{calcVar}()},
\code{\link{estimate}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{estimation functions}
