% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{tooManyDaysHunted}
\alias{tooManyDaysHunted}
\title{Find records with too many days hunted}
\usage{
tooManyDaysHunted(season_df, species)
}
\arguments{
\item{season_df}{Season data tibble}

\item{species}{Species abbreviation, may be one of: 'WF', 'DV', 'SCRG', 'WK', or 'CR'.}
}
\description{
Internal function used in \code{\link{checkSeason}}. Find season records with too many days hunted.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
