% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{tooManyDaysHuntedCR}
\alias{tooManyDaysHuntedCR}
\title{Find records with too many days hunting cranes}
\usage{
tooManyDaysHuntedCR(crtotals_df)
}
\arguments{
\item{crtotals_df}{Crane data}
}
\description{
Internal function used in the \code{checkSeason} family of functions. Find
season records with too many days hunted for cranes.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
