% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangleRef}
\alias{wrangleRef}
\title{Wrangle reference data}
\usage{
wrangleRef(ref_data)
}
\arguments{
\item{ref_data}{Reference data}
}
\description{
Internal function used in \code{\link{wrangleMaxBag}} and \code{\link{wrangleDates}}, which does initial data cleaning.
}
\references{
\url{https://github.com/USFWS/HSestimate}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
